+++ Changelog.md
@@ -3,6 +3,12 @@
 - bugfix save load DSO01 / DSO02 model : didn't do anything
 - extending messages to main window
 - adding imaging popup for use with ASCOM Cameras
+- image widget with zoom and strech for pictures
+- changed modeling behaviour for all: first base then refine in separate steps
+- auto save base model and refinement model in mount und referenced names
+- added color schemes for image view
+- added automatic image show, if ascom camera is used
+- model point could be generated without having a camera connected
 
 2.1.7
 - bugfix